requirements.txt
lightning==1.8.6
torchvision==0.14.0
onnxruntime==1.13.1